backend/baadae/requirements.txt
djangorestframework
djangorestframework-simplejwt
pillow
